requirements.txt
asgiref==3.7.2
boto3==1.34.74
botocore==1.34.74
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
django-dotenv==1.4.2
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
jmespath==1.0.1
pillow==10.2.0
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
urllib3==2.2.1
